permission for liquidation

+++ src/app/views/businesses/stores/liquidation-materials/liquidation-material-list/liquidation-material-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, HostListener } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { NzModalService } from 'ng-zorro-antd';
 
+import { RoleService } from 'src/app/shared/services/role.service';
 import { NotifyService } from 'src/app/shared/services/notify.service';
 import { LiquidationMaterialService } from 'src/app/shared/services/liquidation-material.service';
 
@@ -39,13 +40,21 @@ export class LiquidationMaterialListComponent implements OnInit {
   @HostListener('window:keydown', ['$event'])
   onKeyPress($event: KeyboardEvent) {
     if (($event.ctrlKey || $event.metaKey) && ($event.keyCode === 73 || $event.keyCode === 105)) {
-      this.router.navigate(['/nghiep-vu/kho/thanh-ly-vat-tu/tao-phieu-thanh-ly']);
+      this.roleService.checkPermission('THANH_LY_VAT_TU', 'Create')
+        .subscribe((response: boolean) => {
+          if (response) {
+            this.router.navigate(['/nghiep-vu/kho/thanh-ly-vat-tu/tao-phieu-thanh-ly']);
+          } else {
+            this.notify.warning('Bạn không có quyền');
+          }
+        });
     }
   }
 
   constructor(private route: ActivatedRoute,
     private router: Router,
     private modalService: NzModalService,
+    private roleService: RoleService,
     private liquidationService: LiquidationMaterialService,
     private notify: NotifyService) { }
 
@@ -73,13 +82,32 @@ export class LiquidationMaterialListComponent implements OnInit {
         this.loading = false;
         this.pagination = res.pagination;
         this.dataSet = res.result;
-      }, error => {
-        this.notify.error('Có lỗi xảy ra');
-        console.log('error getAllPagingLiquidationMaterial');
-      }, () => {
+
         if (this.dataSet.length === 0 && this.pagination.currentPage !== 1) {
           this.pagination.currentPage -= 1;
           this.loadData();
+        }
+      });
+  }
+
+  addNew() {
+    this.roleService.checkPermission('THANH_LY_VAT_TU', 'Create')
+      .subscribe((response: boolean) => {
+        if (response) {
+          this.router.navigate(['/nghiep-vu/kho/thanh-ly-vat-tu/tao-phieu-thanh-ly']);
+        } else {
+          this.notify.warning('Bạn không có quyền');
+        }
+      });
+  }
+
+  update(id: number) {
+    this.roleService.checkPermission('THANH_LY_VAT_TU', 'Update')
+      .subscribe((response: boolean) => {
+        if (response) {
+          this.router.navigate(['/nghiep-vu/kho/thanh-ly-vat-tu/sua-phieu-thanh-ly', id]);
+        } else {
+          this.notify.warning('Bạn không có quyền');
         }
       });
   }
@@ -107,16 +135,21 @@ export class LiquidationMaterialListComponent implements OnInit {
   }
 
   delete(id: number) {
-    this.notify.confirm('Bạn có chắc chắn muốn xóa không?', () => {
-      this.liquidationService.delete(id).subscribe((res: boolean) => {
-        if (res) {
-          this.loadData();
-          this.notify.success('Xóa thành công');
+    this.roleService.checkPermission('THANH_LY_VAT_TU', 'Delete')
+      .subscribe((response: boolean) => {
+        if (response) {
+          this.notify.confirm('Bạn có chắc chắn muốn xóa không?', () => {
+            this.liquidationService.delete(id).subscribe((res: boolean) => {
+              if (res) {
+                this.loadData();
+                this.notify.success('Xóa thành công');
+              }
+            });
+          });
+        } else {
+          this.notify.warning('Bạn không có quyền');
         }
-      }, error => {
-        console.log('error deleteLiquidationMaterial');
       });
-    });
   }
 
   search() {
